refactor: js파일명 소문자로 수정
